requirements.txt
apscheduler==3.10.4
black==24.4.2
celery==5.4.0
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.0.0
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
flower==2.0.1
humanize==4.9.0
mccabe==0.7.0
packaging==24.1
pillow==10.3.0
python-telegram-bot==20.0
psycopg2-binary==2.9.9
redis==5.0.6
stripe==9.11.0
tornado==6.4.1
tzdata==2024.1
dj-database-url==2.2.0

Django~=5.0.6
requests~=2.32.3